WIP: 레이어드 아키텍처 개선 - DTO 중첩 클래스 변경 및 역방향 의존성 제거

+++ src/main/kotlin/kr/hhplus/be/server/application/couponevent/CouponEventCriteria.kt
@@ -26,7 +26,7 @@ sealed class CouponEventCriteria {
     data class Create(
         val benefitMethod: String,
         val benefitAmount: String,
-        val totalIssueAmount: Int
+        val totalIssueAmount: Long
     ) : CouponEventCriteria() {
         init {
             require(benefitMethod.isNotBlank()) { "혜택 방식은 비어있을 수 없습니다." }
