+++ src/main/kotlin/kr/hhplus/be/server/controller/couponevent/CouponEventResponse.kt
@@ -11,8 +11,8 @@ sealed class CouponEventResponse {
         val id: String,
         val benefitMethod: String,
         val benefitAmount: String,
-        val totalIssueAmount: Int,
-        val leftIssueAmount: Int
+        val totalIssueAmount: Long,
+        val leftIssueAmount: Long
     ) : CouponEventResponse()
     
     /**
